Fix bindgen enum signedness breaking Linux/macOS builds
Error::UnsupportedFeature held an i32 and converter.rs passed
option.kind() (sys::dxil_spv_option) directly. bindgen generates that
enum as c_int on Windows but c_uint on Linux/macOS (the upstream header
has no negative enumerators, so clang picks an unsigned repr), so the
add_option() error path failed to compile off Windows with E0308
mismatched types.

Normalize the safe-layer payload to u32 (option tags are always
non-negative) and cast at the call site, which is portable across both
bindgen representations.

+++ dxil-spirv/src/converter.rs
@@ -130,7 +130,7 @@ impl Converter {
     /// by the linked library.
     pub fn add_option(&mut self, option: &ConverterOption) -> Result<()> {
         if !option.is_supported() {
-            return Err(Error::UnsupportedFeature(option.kind()));
+            return Err(Error::UnsupportedFeature(option.kind() as u32));
         }
         let (_base, data) = option.to_raw();
         let base = data.as_base();
